feat: can render the top hackernews story titles.
The "Lister" component will form the basis of a given list of items on hackernews.

It takes advantage of lit-html's helpers for rendering lists of data, as well as handling of lazily loaded information via promises.

Its responsibilities aren't well organised right now; the component and its supporting data will be refactored to allow easier maintenance.