wip: create my own routing components; extract routing to there.
why? because the sk-router does not correctly pass properties to the components it hydrates. I suppose I should actually just fix that.

This is just an extraction of what I had before. It doesn't work yet because I've not made any mapping between routes registered and how they are rendered.

I don't actually like the component hierarchy I've devised here. It's a poor mirror of react-router where each route is its own component. It's probably better to built it more like vue-router, where you have a plain old JS map of routes to components, and a single component to render the current route. This lends itself to the PWA style far more.