refactor: define routes separately to the client; rewrite routes and route handling logic for the client-side.
The original thought here was that by creating a single definition of routes for the whole application and wrapping those in the client and server as appropriate, it would make understanding the app easier. In practice, I'm not sure that abstracting the route paths and names has helped much... the handling of routes still requires a different map on client and server, so the handling gets less obvious rather than moreso.