fix: handle server error content and render more types of errors when they occur.
This is definitely not the most elegant of error handling solutions. Indeed, it seems quite convoluted. Still, it is helping to propagate messages from the server to the UI, which is nice.

Note HTTP status code 0 is likely when network requests time out, hosts cannot be resolved, or generally a response never returns for a given request.