chore: cleanup and improvement of webpack config.
Registering the serviceworker source as a chunk means that webpack will watch it for changes and re-compile in development mode. This means we don't need to stop and start webpack every time the serviceworker changes.

Of course, registering this as a chunk means it gets compiled and output everywhere our other application code would be, too, so we need to exclude it from some places.