fix: the service worker now responds to application routes.
In the previous arrangement, the API routes were not being cached whatsoever, because the regex was matching against the full href of the URL rather than its path.

There's an odd incongruity, though: if you register regexes for the whitelist or blacklist of a navigation route, these match against the path + search of the URL, not the full href.

This seems like something that could be improved in workbox.