feat: unpack and seed cache with items from requests to list endpoints.
In a naive REST implementation, the list endpoint would return the IDs of items to make subsequent requests for.
Our REST endpoint for lists returns a collection of items, which makes it not really RESTful.
This may indicate that the REST design needs some re-thinking.

An alternative design might see us do away with the REST requests altogether and adopt websockets, so we could keep a persistent connection to the server and reduce the cost of making N+1 requests for items.

Meanwhile, in the current reality where REST is our transfer mechanism and we care about providing data to users quickly, unpacking data from a given response to seed a separate cache seems like a reasonable compromise.